T#766: Extract governance routes to src/governance/routes.ts
Phase 1.1 of Library #102 modularization. Moves 12 route handlers + decorateRule helper
from server.ts to src/governance/routes.ts. Table creation and seed data stay in server.ts.
No logic changes — code moved as-is.
